+++ swift/tuners/adapter.py
@@ -154,7 +154,7 @@ class Adapter:
         modules: List[torch.nn.Module] = find_sub_module(
             module, f'adapter_{adapter_name}')
         for _module in modules:
-            _module.activate(activate)
+            _module.set_activation(activate)
 
 
 class AdapterModule(nn.Module):
@@ -195,7 +195,7 @@ class AdapterModule(nn.Module):
 
         self.apply(_init_weights)
 
-    def activate(self, activate=True):
+    def set_activation(self, activate=True):
         self._activate = activate
 
     def forward(self, x, identity=None):
